+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Added trigger `Lead Status Change`
 * Added ability to pass trigger field to `Updated Lead`, `Updated Contact`, `Updated Deal`. User will be able to select field which value will be validated against trigger - Zap won't trigger if value doesnt' change
+* Support for 429 responses from Public API (Rate Limiting) - App will not break while trying to parse response
 
 ## 1.0.13